Bone, Node, VertexWeight in proper classes, material binding.
