+++ game-engine/src/main/java/gameengine/engine/asset/Skeleton.java
@@ -2,7 +2,33 @@ package gameengine.engine.asset;
 
 import java.util.List;
 
+import org.joml.Matrix4f;
+
 public class Skeleton {
+	public static class Bone {
+		private int boneID;
+		private String boneName;
+		private Matrix4f offsetTransform;
+		
+		public Bone(int boneID, String boneName, Matrix4f offsetTransform) {
+			this.boneID = boneID;
+			this.boneName = boneName;
+			this.offsetTransform = offsetTransform;
+		}
+		
+		
+		public int getID() {
+			return this.boneID;
+		}
+		
+		public String getName() {
+			return this.boneName;
+		}
+		
+		public Matrix4f getOffsetTransform() {
+			return this.offsetTransform;
+		}
+	}
 
 	private List<Bone> bones;
